Commented out KFSBasePage#expected_element
In the long run, we should be able to uncomment the line because anytime we try to go to a KFSBasePage this element should exist. Unfortunately, we can't for now because if you use on(KFSBasePage), execution will hang until the expected_element materializes (or it'll error out). That's to be expected, but we have some checks that try to use on(KFSBasePage) to determine whether we are actually *on* a given page (see AssetManualPaymentObject#build). In the case that we aren't, these checks don't work as intended.